feat: add confirm delete modal on Die table

+++ application/react/Die/DieTable/DieRowActions/DieRowActions.tsx
@@ -10,14 +10,17 @@ import { IoCreateOutline } from "react-icons/io5";
 import { IoTrashOutline } from "react-icons/io5";
 import { IDie } from '@shared/types/models';
 import { MongooseId } from '@shared/types/typeAliases';
+import { ConfirmationResult } from '../../../_global/Modal/useConfirmation';
 
 type Props = {
   row: Row<IDie>
+  confirmation: ConfirmationResult
 }
 
 export const DieRowActions = (props: Props) => {
-  const { row } = props;
-  const { _id : mongooseObjectId } = row.original;
+  const { row, confirmation } = props;
+  const { _id: mongooseObjectId } = row.original;
+  const { showConfirmation } = confirmation;
 
   const navigate = useNavigate();
   const queryClient = useQueryClient()
@@ -25,12 +28,19 @@ export const DieRowActions = (props: Props) => {
   const onEditClicked = (mongooseObjectId: MongooseId) => {
     navigate(`/react-ui/forms/die/${mongooseObjectId}`)
   }
-  
-  const onDeleteClicked = (mongooseObjectId: MongooseId) => {
-    alert('@TODO Storm: Add a confirmation modal before deletion?')
+
+  const onDeleteClicked = async (mongooseObjectId: MongooseId) => {
+    const confirmed = await showConfirmation({
+      title: 'Delete Die',
+      message: 'Are you sure you want to delete this die? This action cannot be undone.',
+      confirmText: 'Delete',
+    });
+
+    if (!confirmed) return;
+
     axios.delete(`/dies/${mongooseObjectId}`)
-      .then((_ : AxiosResponse) => {
-        queryClient.invalidateQueries({ queryKey: [GET_DIES_QUERY_KEY]})
+      .then((_: AxiosResponse) => {
+        queryClient.invalidateQueries({ queryKey: [GET_DIES_QUERY_KEY] })
         useSuccessMessage('Deletion was successful')
       })
       .catch((error: AxiosError) => useErrorMessage(error))
@@ -40,6 +50,6 @@ export const DieRowActions = (props: Props) => {
     <RowActions>
       <RowActionItem text='Edit' Icon={IoCreateOutline} onClick={() => onEditClicked(mongooseObjectId)} />
       <RowActionItem text='Delete' Icon={IoTrashOutline} onClick={() => onDeleteClicked(mongooseObjectId)} />
-  </RowActions>
+    </RowActions>
   )
 }
